AlertSliderPlugin: Allow null navigation bar in setup
The navBar parameter in the setup method was previously assumed to be non-null. This commit updates the method signature to allow navBar to be null. This change prevents a potential NullPointerException that could occur if the navigation bar was not provided during the setup of the overlay plugin (e.g., in AlertSliderPlugin). Null checks have been added to handle cases where navBar is null.

AndroidRuntime: java.lang.NullPointerException: Parameter specified as non-null is null: method org.lineageos.settings.device.AlertSliderPlugin.setup, parameter navBar

Signed-off-by: QuinceROMs <quinceroms@gmail.com>